Add a HasSystemProperties trait and use it
Incidentally fixes #101
