+++ src/Delivery/Synchronization/DeletedResource.php
@@ -6,6 +6,7 @@
 
 namespace Contentful\Delivery\Synchronization;
 
+use Contentful\Delivery\HasSystemProperties;
 use Contentful\Delivery\SystemProperties;
 
 /**
@@ -13,10 +14,7 @@ use Contentful\Delivery\SystemProperties;
  */
 abstract class DeletedResource implements \JsonSerializable
 {
-    /**
-     * @var SystemProperties
-     */
-    private $sys;
+    use HasSystemProperties;
 
     /**
      * DeletedResource constructor.
@@ -26,78 +24,6 @@ abstract class DeletedResource implements \JsonSerializable
     public function __construct(SystemProperties $sys)
     {
         $this->sys = $sys;
-    }
-
-    /**
-     * Returns the ID of the resource.
-     *
-     * @return string
-     *
-     * @api
-     */
-    public function getId()
-    {
-        return $this->sys->getId();
-    }
-
-    /**
-     * Returns the space the resource used to belong to.
-     *
-     * @return \Contentful\Delivery\Space
-     *
-     * @api
-     */
-    public function getSpace()
-    {
-        return $this->sys->getSpace();
-    }
-
-    /**
-     * Returns the last revision of the resource before it was deleted.
-     *
-     * @return int
-     *
-     * @api
-     */
-    public function getRevision()
-    {
-        return $this->sys->getRevision();
-    }
-
-    /**
-     * Returns the time when the resource was updated.
-     *
-     * @return \DateTimeImmutable
-     *
-     * @api
-     */
-    public function getUpdatedAt()
-    {
-        return $this->sys->getUpdatedAt();
-    }
-
-    /**
-     * Returns the time when the resource was created.
-     *
-     * @return \DateTimeImmutable
-     *
-     * @api
-     */
-    public function getCreatedAt()
-    {
-        return $this->sys->getCreatedAt();
-    }
-
-    /**
-     * Returns the time when the resource was deleted.
-     *
-     * @return \DateTimeImmutable
-     *
-     * @api
-     */
-    public function getDeletedAt()
-    {
-        return $this->sys->getDeletedAt();
     }
 
     /**
